requirements.txt
setuptools==57.5.0
aiofiles==0.5.0
amqp==1.4.9
aniso8601==7.0.0
anyio==3.3.0
anyjson==0.3.3
appdirs==1.4.3
arrow==1.1.1
asgiref==3.3.1
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.1.0
bcrypt==3.2.0
billiard==3.3.0.23
binaryornot==0.4.4
blessed==1.17.9
boto3==1.24.69
botocore==1.27.69
CacheControl==0.12.6
cached-property==1.5.1
cachetools==4.1.1
celery==3.1.26.post2
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cookiecutter==1.7.3
cryptography==37.0.4
DAWG-Python==0.7.2
defusedxml==0.6.0
Deprecated==1.2.13
distlib==0.3.0
distro==1.7.0
Django==3.1.6
django-allauth==0.42.0
django-celery==3.3.1
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cors-headers==3.4.0
django-front==0.5.13
django-health-check==3.12.1
django-inlineedit==1.2
django-inplaceedit==1.4.1
django-js-asset==1.2.2
django-sql-explorer==2.1.2
django-storages==1.9.1
django-telegram-login==0.2.3
dnspython==2.1.0
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
email-validator==1.1.3
fastapi==0.63.0
future==0.16.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.3
grpcio==1.26.0
gunicorn
h11==0.12.0
html5lib==1.0.1
httpcore==0.13.6
httplib2==0.18.1
httptools==0.1.2
httpx==0.18.2
idna==3.3
importlib-metadata==3.10.1
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.3.1
kombu==3.0.37
lockfile==0.12.2
Markdown==3.3
MarkupSafe==2.0.1
mock==4.0.2
msgpack==0.6.2
mysqlclient==2.0.1
netaddr==0.8.0
oauthlib==3.1.0
packaging==20.3
paramiko==2.7.1
pathspec==0.5.9
pep517==0.8.2
Pillow==8.2.0
poyo==0.5.0
progress==1.5
promise==2.3
protobuf==3.12.4
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pydantic-sqlalchemy==0.0.8.post1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pysqlite3==0.5.3
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.0
python-dotenv==0.17.1
python-multipart==0.0.5
python-slugify==5.0.2
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
requests==2.28.1
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.5.0
rsa==4.6
Rx==1.6.1
s3transfer==0.6.0
semantic-version==2.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.17
sqlparse==0.4.1
starlette==0.13.6
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.2
toml==0.10.2
typing-extensions==3.10.0.1
ujson==5.10.0
uritemplate==3.0.1
urllib3==1.26.12
uvicorn==0.13.4
watchgod==0.7
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
wrapt==1.14.1
yet-another-django-profiler==1.1.0
yookassa==2.3.1
zipp==3.8.1
